Fix compilation with older Android NDK / API level
Using an older Android target API level and NDK, `log2f` appears to not work properly. Using `HAVE_LOG2F` results in a linker error that the symbol for "log2f" cannot be found, but compiling without results in a conflict due to `static inline` vs. `inline` in the standard library.

This is one solution for resolving compilation for older Android NDK environments