Move BackendListEnd variable from global scope to function scope.
It reduces the scope of this variable to the function it is used, which
is cleaner.

It also avoid using the variable before it has been initialized.
Cf  StackOverflow: https://stackoverflow.com/a/25009120/5112390
~~~
The standard says that all objects are initialized in the same
translation unit (aka object file which corresponds to a single source
file) before any function is called in that translation unit.

The standard is allowing lazy initialization to occur in the event that
a DLL is loaded at run time. If you allow run-time linkage of your code,
you can't say that everything is initialized before main().
~~~

Bug: https://github.com/kcat/openal-soft/issues/288