Don't pack ambdec coefficients
Tne coefficients are placed as for full 3D ACN handling. The ChanMask just
indicates which have potentially useful values. This could be a bit more
agressive and clear ChanMask bits for channels that don't contribute to output,
so that a decoder the specifies height-related channel bits, but leaves their
coefficients all 0, can be handled as 2D. I don't expect many ambdec files to
be like that, though.