Disable MSVC warning C5051
"attribute 'attribute-name' requires at least 'standard-level'; ignored"

This hides the fact that it's not optimizing [[likely]] and [[unlikely]], even
though it could (GCC does, Clang gives a similar warning to MSVC, but is
ambiguous about whether it's still doing the optimization). There's nothing I
know to do to make MSVC do those optimizations it otherwise knows to do, so
just hide the warnings instead of pessimizing the code.