Use the object serial ID from PipeWire when available
Using the node ID as a target is deprecated in newer versions of PipeWire. The
serial ID is a monotonic 64-bit integer ID, incremeneted for every object
created, so is guaranteed to always refer to the same target (until it wraps
around, which I suppose isn't expected/allowed to happen), compared to the
32-bit node ID which I guess allows reuse.

We could instead use the target node's name instead of the serial ID, but an
integer is nicer to manage than a string.