Simplify committing EAX effect properties
There's no need to explicitly clamp to EFX limits when they're the same as or
more lenient than EAX, which were already validated when set, or when it's
within tolerance of the effect implementation.

Also it's generally better to check once all properties for changes and apply
them all if one is different, rather than checking and setting each member
individually.